Consolidate deployment checking functions and move the test functions into test files
