+++ test/e2e/knativeservingdeployment_test.go
@@ -18,6 +18,9 @@ package e2e
 import (
 	"testing"
 
+	apierrs "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/wait"
 	"knative.dev/pkg/test/logstream"
 	"knative.dev/serving-operator/test"
 	"knative.dev/serving-operator/test/resources"
@@ -58,4 +61,85 @@ func TestKnativeServingDeployment(t *testing.T) {
 		knativeServingVerify(t, clients, names)
 		knativeServingDeletion(t, clients, names)
 	})
+}
+
+// knativeServingVerify verifies if the KnativeServing can reach the READY status.
+func knativeServingVerify(t *testing.T, clients *test.Clients, names test.ResourceNames) {
+	if _, err := resources.WaitKnativeServingReady(clients.KnativeServingAlphaClient, names.KnativeServing,
+		resources.IsKnativeServingReady); err != nil {
+		t.Fatalf("KnativeService %q failed to get to the READY status: %v", names.KnativeServing, err)
+	}
+
+}
+
+// deploymentRecreation verify whether all the deployments for knative serving are able to recreate, when they are deleted.
+func deploymentRecreation(t *testing.T, clients *test.Clients, names test.ResourceNames) {
+	dpListSave, err := clients.KubeClient.Kube.AppsV1().Deployments(names.Namespace).List(metav1.ListOptions{})
+	if err != nil {
+		t.Fatalf("Failed to get any deployment under the namespace %q: %v",
+			test.ServingOperatorNamespace, err)
+	}
+	if len(dpListSave.Items) == 0 {
+		t.Fatalf("No deployment under the namespace %q was found",
+			test.ServingOperatorNamespace)
+	}
+	// Delete the deployments one by one to see if they will be recreated.
+	for _, deployment := range dpListSave.Items {
+		if err := clients.KubeClient.Kube.AppsV1().Deployments(deployment.Namespace).Delete(deployment.Name,
+			&metav1.DeleteOptions{}); err != nil {
+			t.Fatalf("Failed to delete deployment %s/%s: %v", deployment.Namespace, deployment.Name, err)
+		}
+
+		waitErr := wait.PollImmediate(resources.Interval, resources.Timeout, func() (bool, error) {
+			dep, err := clients.KubeClient.Kube.AppsV1().Deployments(deployment.Namespace).Get(deployment.Name, metav1.GetOptions{})
+			if err != nil {
+				// If the deployment is not found, we continue to wait for the availability.
+				if apierrs.IsNotFound(err) {
+					return false, nil
+				}
+				return false, err
+			}
+			return resources.IsDeploymentAvailable(dep)
+		})
+
+		if waitErr != nil {
+			t.Fatalf("The deployment %s/%s failed to reach the desired state: %v", deployment.Namespace, deployment.Name, err)
+		}
+
+		if _, err := resources.WaitKnativeServingReady(clients.KnativeServingAlphaClient, test.ServingOperatorName,
+			resources.IsKnativeServingReady); err != nil {
+			t.Fatalf("KnativeService %q failed to reach the desired state: %v", test.ServingOperatorName, err)
+		}
+		t.Logf("The deployment %s/%s reached the desired state.", deployment.Namespace, deployment.Name)
+	}
+}
+
+// knativeServingDeletion deletes tha KnativeServing to see if all the deployments will be removed.
+func knativeServingDeletion(t *testing.T, clients *test.Clients, names test.ResourceNames) {
+	if err := clients.KnativeServingAlphaClient.Delete(names.KnativeServing, &metav1.DeleteOptions{}); err != nil {
+		t.Fatalf("KnativeService %q failed to delete: %v", names.KnativeServing, err)
+	}
+
+	dpListSave, err := clients.KubeClient.Kube.AppsV1().Deployments(names.Namespace).List(metav1.ListOptions{})
+	if err != nil {
+		t.Fatalf("Error getting any deployment under the namespace %q: %v", names.Namespace, err)
+	}
+
+	for _, deployment := range dpListSave.Items {
+		waitErr := wait.PollImmediate(resources.Interval, resources.Timeout, func() (bool, error) {
+			_, err := clients.KubeClient.Kube.AppsV1().Deployments(deployment.Namespace).Get(deployment.Name, metav1.GetOptions{})
+			if err != nil {
+				if apierrs.IsNotFound(err) {
+					return true, nil
+				}
+				return false, err
+			}
+			return false, nil
+		})
+
+		if waitErr != nil {
+			t.Fatalf("The deployment %s/%s failed to be deleted: %v", deployment.Namespace, deployment.Name, err)
+		}
+		t.Logf("The deployment %s/%s has been deleted.", deployment.Namespace, deployment.Name)
+	}
 }
